# httpserver/dto/user_dto.go
package dto

type UpsertUserDto struct {
	Email    string `json:"email" binding:"required"`
	Password string `json:"password" binding:"required"`
	FullName string `json:"full_name" binding:"required"`
	Role     string `json:"role" binding:"required"`
}

type LoginDto struct {
	Email    string `json:"email" binding:"required"`
	Password string `json:"password" binding:"required"`
}

type TopUpBalanceDto struct {
	Balance uint `json:"balance" binding:"required,gt=0,lte=1000000000"`
}